% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cjsg_trf3.R
\name{baixar_cjsg_trf3}
\alias{baixar_cjsg_trf3}
\title{Baixa decisões de segunda instância do TRF3}
\usage{
baixar_cjsg_trf3(
  livre = "",
  aspas = FALSE,
  data_inicial = "",
  data_final = "",
  relator = "0",
  classe = "0",
  orgao_julgador = "0",
  diretorio = "."
)
}
\arguments{
\item{livre}{busca livre}

\item{aspas}{TRUE para colocar a busca entre aspas}

\item{data_inicial}{formato "dd/mm/aaaa"}

\item{data_final}{formato "dd/mm/aaaa"}

\item{relator}{Obtenha a lista completa com a função trf3_magistrados}

\item{classe}{Obtenha a lista completa com a função trf3_classes}

\item{orgao_julgador}{Obtenha a lista completa com a função trf3_orgao_julgador}

\item{diretorio}{Default para atual.}
}
\value{
Baixa os htmls nos diretório especificado
}
\description{
Baixa decisões de segunda instância do TRF3
}
\examples{
\dontrun{
baixar_cjsg_trf3(
  livre = "agência nacional",  data_inicial = "10/07/2019",
  data_final = "31/07/2019"
)
}
}
