% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cposg_trf3.R
\name{baixar_cposg_trf3}
\alias{baixar_cposg_trf3}
\title{Baixar metadados das decis천es de segundo grau}
\usage{
baixar_cposg_trf3(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Vetor de processos}

\item{diretorio}{Diret처rio onde ser찾o armazenados os processos}
}
\value{
htmls
}
\description{
Baixar metadados das decis천es de segundo grau
}
