% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_trf3_ordem.R
\name{baixar_trf3_ordem}
\alias{baixar_trf3_ordem}
\title{Baixar julgados do trf3 por ordem}
\usage{
baixar_trf3_ordem(ordem, diretorio)
}
\arguments{
\item{ordem}{vetor com números de ordem}

\item{diretorio}{diretório}
}
\value{
baixa html no destino designado
}
\description{
Baixar julgados do trf3 por ordem
}
