% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_na}
\alias{is_na}
\title{Função usada para selecionar apenas variáveis com algum valor}
\usage{
is_na(x)
}
\arguments{
\item{x}{variaveis}
}
\description{
Função usada para selecionar apenas variáveis com algum valor
}
