% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cpopg_trf3.R
\name{ler_cpopg_trf3}
\alias{ler_cpopg_trf3}
\title{Lê processos de primeiro grau do TRF3}
\usage{
ler_cpopg_trf3(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos. Se NULL, informar diretório}

\item{diretorio}{Informar diretório se arquivos = NULL}
}
\value{
tibble
}
\description{
Lê processos de primeiro grau do TRF3
}
