% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cposg_trf3.R
\name{ler_cposg_trf3}
\alias{ler_cposg_trf3}
\title{Lê metadados processuais}
\usage{
ler_cposg_trf3(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informados os arquivos}
}
\value{
tibble
}
\description{
Lê metadados processuais
}
