% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao_trf3.R
\name{ler_movimentacao_trf3}
\alias{ler_movimentacao_trf3}
\title{Ler movimentacao do TRF3}
\usage{
ler_movimentacao_trf3(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos. Se NULL, informar diretório.}

\item{diretorio}{Diretório se arquivos não informados.}
}
\value{
tibble
}
\description{
Ler movimentacao do TRF3
}
\examples{
\dontrun{
df <- ler_movimentacao_trf3()
}
}
