% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes_cjsg_trf3.R
\name{ler_partes_cjsg_trf3}
\alias{ler_partes_cjsg_trf3}
\title{Lê partes do processo de segunda instância do trf3}
\usage{
ler_partes_cjsg_trf3(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de caminhos para os artigos}

\item{diretorio}{Se arquivos não forem fornecidos, fornecer
diretório}
}
\value{
data.frame
}
\description{
Lê partes do processo de segunda instância do trf3
}
\examples{
\dontrun{
ler_partes_cjsg_trf3(diretorio = ".")
}
}
