% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_cjsg_trf3.R
\name{organizar_cjsg_trf3}
\alias{organizar_cjsg_trf3}
\title{Organiza cjsg do TRF3}
\usage{
organizar_cjsg_trf3(df, julgado)
}
\arguments{
\item{df}{dados}

\item{julgado}{coluna onde se encontram os julgados}
}
\value{
mesma tibble acrescida data de julgamento, vara e relator
}
\description{
Organiza cjsg do TRF3
}
\details{
Esta função é temporária, seus termos já foram incluídos
    na função ler_dados_cjsg_trf3
}
\examples{
\dontrun{
df <- organizar_cjsg_trf3(df, julgado)
}
}
