% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_cpopg_trf3.R
\name{organizar_cpopg_trf3}
\alias{organizar_cpopg_trf3}
\title{Cria tibbles dados e partes organizados}
\usage{
organizar_cpopg_trf3(df)
}
\arguments{
\item{df}{Lido com ler_cpopg_trf3}
}
\value{
Dois tibbles: dados_cpopg_trf3 e partes_cpopg_trf3
}
\description{
Cria tibbles dados e partes organizados
}
