% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_add_fkey.R
\name{psql_add_fkey}
\alias{psql_add_fkey}
\title{Adiciona foreign key}
\usage{
psql_add_fkey(con, mtbl, dtbl, constraint, key = "processo")
}
\arguments{
\item{con}{conexao}

\item{mtbl}{tabela principal}

\item{dtbl}{tabela a receber a chave estrangeiro}

\item{constraint}{nome do constraint}

\item{key}{variável de referência}
}
\value{
0
}
\description{
Adiciona foreign key
}
\examples{
\dontrun{
psql_add_fkey(con, "dados", "partes",constraint = "dados_parte", key = "ordem")
}
}
