% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_distinct.R
\name{psql_distinct}
\alias{psql_distinct}
\title{Elimina duplicados}
\usage{
psql_distinct(con, tbl = NULL, pkey = NULL, coluna = "ordem")
}
\arguments{
\item{con}{conexao}

\item{tbl}{tabela}

\item{pkey}{chave primária}

\item{coluna}{nome da coluna de referência}
}
\value{
nada
}
\description{
Elimina duplicados
}
