% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_trf3_query.R
\name{psql_trf3_query}
\alias{psql_trf3_query}
\title{PostgreSQL full-text search}
\usage{
psql_trf3_query(
  con,
  tbl,
  query = "",
  classes = NULL,
  processos = NULL,
  origem = c("SP", "MS"),
  start = "2009-01-01",
  end = NULL,
  partes = "partes"
)
}
\arguments{
\item{con}{connection}

\item{tbl}{table}

\item{query}{palavras a serem buscadas}

\item{classes}{vetor de classes}

\item{processos}{vetor de processos}

\item{origem}{vetor de origem}

\item{start}{data inicial no formato "yyyy-mm-dd"}

\item{end}{data final no formato "yyyy-mm-dd"}

\item{partes}{nome da base partes}
}
\value{
tibble
}
\description{
PostgreSQL full-text search
}
\examples{
\dontrun{
dplyr::copy_to(con, "julgados", df)
df <- psql_trf3_query(con, "julgados", "IRPF")
}
}
