% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_trf3_sample.R
\name{psql_trf3_sample}
\alias{psql_trf3_sample}
\title{Gets a table's sample}
\usage{
psql_trf3_sample(con, tbl = NULL, column = NULL, value = NULL,
  n = 10)
}
\arguments{
\item{con}{connection}

\item{tbl}{table name}

\item{column}{optional column name to apply filter}

\item{value}{optional column value to apply filter}

\item{n}{number of rows, defaults to 10.}
}
\value{
a data.frame sample according to specified arguements.
}
\description{
Gets a table's sample
}
