% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_trf3_tokenize.R
\name{psql_trf3_tokenize}
\alias{psql_trf3_tokenize}
\title{Cria um index a partir dos julgados}
\usage{
psql_trf3_tokenize(con, tbl)
}
\arguments{
\item{con}{conexão}

\item{tbl}{tabela}
}
\value{
Nenhum objeto R é retornado
}
\description{
Cria um index a partir dos julgados
}
\examples{
\dontrun{
library(carf)
con <- dbx::dbxConnect()
dplyr::copy_to(con, "julgados", df)
psql_trf3_tokenize(con, "julgados")
}
}
