% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_trf3_trigger.R
\name{psql_trf3_trigger}
\alias{psql_trf3_trigger}
\title{Adds trigger to cjpg table}
\usage{
psql_trf3_trigger(con, tbl, config = "pg_catalog.portuguese")
}
\arguments{
\item{con}{connection}

\item{tbl}{table name}

\item{config}{defaults to portuguese}
}
\value{
query result is cleared after trigger creation
}
\description{
This function adds a trigger to a cjpg table retokenize
    the text the same table every time it is updated or new data is
    inserted to it.
}
