% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_write_trf3.R
\name{psql_write_trf3}
\alias{psql_write_trf3}
\title{Cria e salva uma tabela no PostgreSQL}
\usage{
psql_write_trf3(con = NULL, tbl = NULL, data = NULL)
}
\arguments{
\item{con}{a connection}

\item{tbl}{table name}

\item{data}{trf3 data.frame to be written.}
}
\value{
returns NULL if everything went well
}
\description{
Esta função cria uma tabela com base
    nas colunas a serem salvas. Em seguida adiciona
    uma chave primária chamada document_ix. Depois
    disso insere as observações. Ao final, cria um
    índice remissivo chamado document_tokens com
    base no algorítimo GIN.
}
\examples{
\dontrun{
con <- dbx::dbxConnect()
psql_write_trf3(con, "consumidor", consumidor)
}
}
