% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf3_baixar_acordaos.R
\name{trf3_baixar_acordaos}
\alias{trf3_baixar_acordaos}
\title{Baixar acórdãos com número de ordem}
\usage{
trf3_baixar_acordaos(processo, diretorio = ".")
}
\arguments{
\item{processo}{Número do processo}

\item{diretorio}{Diretório}
}
\value{
htmls
}
\description{
Baixar acórdãos com número de ordem
}
