% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf3_distribuidores.R
\docType{data}
\name{trf3_distribuidores}
\alias{trf3_distribuidores}
\title{trf3_distribuidores}
\format{
tibble com 85 linhas e duas variáveis:
\describe{
  \item{\code{secao}}{character seção do TRF3}
  \item{\code{distribuidor}}{character código do distribuidor}
}
}
\source{
\url{https://cnj.jus.br/}
}
\usage{
trf3_distribuidores
}
\description{
Distribuidores do TRF3
}
\keyword{datasets}
