% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf3_ler_cjsg.R
\name{trf3_ler_cjsg}
\alias{trf3_ler_cjsg}
\title{Lê htmls baixados por trf3_baixar_cjsg}
\usage{
trf3_ler_cjsg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se forem informados, diretorio é ignorado.}

\item{diretorio}{Default para o atual}
}
\value{
tibble com informações processuais
}
\description{
Lê htmls baixados por trf3_baixar_cjsg
}
\examples{
\dontrun{
ler_cjsg_trf3(diretorio = ".")
}
}
