% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf3_ler_movimentacao.R
\name{trf3_ler_movimentacao}
\alias{trf3_ler_movimentacao}
\title{Ler movimentacao do TRF3}
\usage{
trf3_ler_movimentacao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos. Se NULL, informar diretório.}

\item{diretorio}{Diretório se arquivos não informados.}
}
\value{
tibble
}
\description{
Ler movimentacao do TRF3
}
\examples{
\dontrun{
df <- trf3_ler_movimentacao()
}
}
