% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf3_orgaos_julgadores.R
\name{trf3_orgaos_julgadores}
\alias{trf3_orgaos_julgadores}
\title{Obtêm lista atualizada dos órgãos julgadores}
\usage{
trf3_orgaos_julgadores()
}
\value{
tibble
}
\description{
Obtêm lista atualizada dos órgãos julgadores
}
