% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjsg_trf3.R
\name{ler_cjsg_trf3}
\alias{ler_cjsg_trf3}
\title{Lê htmls baixados por baixar_cjsg_trf3}
\usage{
ler_cjsg_trf3(diretorio = ".")
}
\arguments{
\item{diretorio}{Default para o atual}
}
\value{
tibble com informações processuais
}
\description{
Lê htmls baixados por baixar_cjsg_trf3
}
\examples{
\dontrun{
ler_cjsg_trf3(diretorio = ".")
}
}
